SaS clustering prototype
float => double
TfIdf featurizer
simple server & html/js

+++ cluster/em_cluster.hpp
@@ -37,7 +37,7 @@ public:
     bpl::object get_sample_probabilities(
         const std::string & sample_uid);
 
-    float iterate();
+    double iterate();
 
 private:
 
@@ -47,8 +47,8 @@ private:
 
         bool is_hard;
 
-        float prob_sample_class;
-        float prob_class_sample;
+        double prob_sample_class;
+        double prob_class_sample;
     };
 
     typedef std::vector<sample_cluster_state_t> sample_cluster_states_t;
@@ -64,7 +64,7 @@ private:
 
     void compute_pclass_norms(
         sample_cluster_states_t & states,
-        float & hard_norm, float & soft_norm);
+        double & hard_norm, double & soft_norm);
 
     std::vector<std::string> _clusters;
     std::vector<estimator::ptr_t> _cluster_estimators;
