+++ cluster/vectorspace_estimator.cpp
@@ -13,13 +13,13 @@ void make_vectorspace_estimator_bindings()
         sparse_vectorspace_estimator::ptr_t,
         bpl::bases<estimator> >(
         "sparse_vectorspace_estimator",
-        bpl::init<const featurizer::ptr_t &, float>());
+        bpl::init<const featurizer::ptr_t &, double>());
 
     bpl::class_<dense_vectorspace_estimator,
         dense_vectorspace_estimator::ptr_t,
         bpl::bases<estimator> >(
         "dense_vectorspace_estimator",
-        bpl::init<const featurizer::ptr_t &, float>());
+        bpl::init<const featurizer::ptr_t &, double>());
 }
 
 };
